+++ .gitignore
@@ -1,5 +1,7 @@
 env
 .vscode
 __pycache__
-.pytest_cache.
+.pytest_cache
 .idea
+*./__pycache
+*/__pycache
