+++ app/tests/user_api_test.py
@@ -1,14 +1,7 @@
 import json
 from flask import url_for
 from app import db
-from app.models import User
-
-from app.tests.test_base import BaseTestCase
-
-
-class TestUserService(BaseTestCase):
-    # some code...
-    pass
+from app.models.user_model import User
 
 
 def test_register(app):
@@ -24,27 +17,23 @@ def test_register(app):
     resp = client.post(url_for('users.register'), data=json.dumps(data),
                        headers={'Content-Type': 'application/json'})
     print(resp.get_json('id'))
+    print(resp)
     u = User.query.get(resp.get_json('id'))
     u.confirm_id = True
     db.session.commit()
     assert resp.status_code == 201
-
-
-def test_login(app):
-    client = app.test_client()
 
     data = {
         "email": "mfsiddique11@gmail.com",
         "password": "Faizan12345"
     }
 
+    client = app.test_client()
     resp = client.post(url_for('users.login'), data=json.dumps(data), headers={'Content-Type': 'application/json'})
     print(resp.data)
 
     assert resp.status_code == 200
 
-
-def test_change_password(app):
     client = app.test_client()
 
     data = {
@@ -57,3 +46,4 @@ def test_change_password(app):
                        headers={'Content-Type': 'application/json'})
     print(resp.data)
     assert resp.status_code == 201
+
